Completed animationcomponent, made player prefab, implemented run and dig anims

+++ Essence/Time.cpp
@@ -7,10 +7,9 @@ void dae::Time::EnterNextFrame()
 {
 	//calculate duration of last frame
 	auto t = std::chrono::high_resolution_clock::now();
-	m_DeltaTime = float(std::chrono::duration_cast<std::chrono::milliseconds>(t - m_PrevFrame).count());
+	m_DeltaTime = double(std::chrono::duration_cast<std::chrono::milliseconds>(t - m_PrevFrame).count());
 	m_FrameTimes.push_back(m_DeltaTime);
 	m_PrevFrame = t;
-
 
 	if (m_FrameTimes.size() > m_AmountOfFramesForAverage) m_FrameTimes.pop_front();
